+++ base/set.jl
@@ -7,6 +7,9 @@ end
 Set() = Set{Any}()
 Set(itr) = Set{eltype(itr)}(itr)
 
+eltype{T}(s::Set{T}) = T
+similar{T}(s::Set{T}) = Set{T}()
+
 function show(io::IO, s::Set)
     print(io,"Set")
     if isempty(s)
@@ -20,36 +23,23 @@ end
 
 isempty(s::Set) = isempty(s.dict)
 length(s::Set)  = length(s.dict)
-eltype{T}(s::Set{T}) = T
-
 in(x, s::Set) = haskey(s.dict, x)
-
 push!(s::Set, x) = (s.dict[x] = nothing; s)
 pop!(s::Set, x) = (pop!(s.dict, x); x)
 pop!(s::Set, x, deflt) = pop!(s.dict, x, deflt) == deflt ? deflt : x
+pop!(s::Set) = (val = s.dict.keys[start(s.dict)]; delete!(s.dict, val); val)
 delete!(s::Set, x) = (delete!(s.dict, x); s)
 
-union!(s::Set, xs) = (for x=xs; push!(s,x); end; s)
-setdiff!(s::Set, xs) = (for x=xs; delete!(s,x); end; s)
-
-similar{T}(s::Set{T}) = Set{T}()
 copy(s::Set) = union!(similar(s), s)
 
 sizehint!(s::Set, newsz) = (sizehint!(s.dict, newsz); s)
 empty!{T}(s::Set{T}) = (empty!(s.dict); s)
-
 rehash!(s::Set) = (rehash!(s.dict); s)
 
 start(s::Set)       = start(s.dict)
 done(s::Set, state) = done(s.dict, state)
 # NOTE: manually optimized to take advantage of Dict representation
 next(s::Set, i)     = (s.dict.keys[i], skip_deleted(s.dict,i+1))
-
-# TODO: simplify me?
-pop!(s::Set) = (val = s.dict.keys[start(s.dict)]; delete!(s.dict, val); val)
-
-join_eltype() = Bottom
-join_eltype(v1, vs...) = typejoin(eltype(v1), join_eltype(vs...))
 
 union() = Set()
 union(s::Set) = copy(s)
@@ -62,6 +52,9 @@ function union(s::Set, sets::Set...)
     return u
 end
 const ∪ = union
+union!(s::Set, xs) = (for x=xs; push!(s,x); end; s)
+join_eltype() = Bottom
+join_eltype(v1, vs...) = typejoin(eltype(v1), join_eltype(vs...))
 
 intersect(s::Set) = copy(s)
 function intersect(s::Set, sets::Set...)
@@ -89,6 +82,7 @@ function setdiff(a::Set, b::Set)
     end
     d
 end
+setdiff!(s::Set, xs) = (for x=xs; delete!(s,x); end; s)
 
 ==(l::Set, r::Set) = (length(l) == length(r)) && (l <= r)
 < (l::Set, r::Set) = (length(l) < length(r)) && (l <= r)
@@ -118,14 +112,6 @@ function unique(C)
     out
 end
 
-function filter!(f::Function, s::Set)
-    for x in s
-        if !f(x)
-            delete!(s, x)
-        end
-    end
-    return s
-end
 function filter(f::Function, s::Set)
     u = similar(s)
     for x in s
@@ -134,4 +120,12 @@ function filter(f::Function, s::Set)
         end
     end
     return u
+end
+function filter!(f::Function, s::Set)
+    for x in s
+        if !f(x)
+            delete!(s, x)
+        end
+    end
+    return s
 end