+++ src/interpreter.c
@@ -171,7 +171,7 @@ static jl_value_t *eval(jl_value_t *e, jl_value_t **locals, size_t nl)
         return e;
     }
     jl_expr_t *ex = (jl_expr_t*)e;
-    jl_value_t **args = &jl_cellref(ex->args,0);
+    jl_value_t **args = jl_array_data(ex->args);
     size_t nargs = jl_array_len(ex->args);
     if (ex->head == call_sym ||  ex->head == call1_sym) {
         if (jl_is_lambda_info(args[0])) {
@@ -447,6 +447,9 @@ static jl_value_t *eval(jl_value_t *e, jl_value_t **locals, size_t nl)
         jl_throw(args[0]);
     }
     else if (ex->head == boundscheck_sym) {
+        return (jl_value_t*)jl_nothing;
+    }
+    else if (ex->head == fastmath_sym) {
         return (jl_value_t*)jl_nothing;
     }
     else if (ex->head == simdloop_sym) {