+++ test/sparse.jl
@@ -155,6 +155,16 @@ for i = 1:5
     @test full(kron(a,b)) == kron(full(a), full(b))
 end
 
+# scale and scale!
+sA = sprandn(3, 7, 0.5)
+dA = full(sA)
+b = randn(7)
+@test scale(dA, b) == scale(sA, b)
+@test scale(dA, b) == scale!(copy(sA), b)
+b = randn(3)
+@test scale(b, dA) == scale(b, sA)
+@test scale(b, dA) == scale!(b, copy(sA))
+
 # reductions
 @test sum(se33)[1] == 3.0
 @test sum(se33, 1) == [1.0 1.0 1.0]
@@ -499,3 +509,53 @@ a = speye(3,5)
 @test size(rot180(a)) == (3,5)
 @test size(rotr90(a)) == (5,3)
 @test size(rotl90(a)) == (5,3)
+
+function test_getindex_algs{Tv,Ti}(A::SparseMatrixCSC{Tv,Ti}, I::AbstractVector, J::AbstractVector, alg::Int)
+    # Sorted vectors for indexing rows.
+    # Similar to getindex_general but without the transpose trick.
+    (m, n) = size(A)
+    minj, maxj = extrema(J)
+    ((I[1] < 1) || (I[end] > m) || (minj < 1) || (maxj > n)) && BoundsError()
+
+    (alg == 0) ? Base.SparseMatrix.getindex_I_sorted_bsearch_A(A, I, J) :
+    (alg == 1) ? Base.SparseMatrix.getindex_I_sorted_bsearch_I(A, I, J) :
+    Base.SparseMatrix.getindex_I_sorted_linear(A, I, J)
+end
+
+let M=2^14, N=2^4
+    Irand = randperm(M);
+    Jrand = randperm(N);
+    SA = [sprand(M, N, d) for d in [1, 0.1, 0.01, 0.001, 0.0001]];
+    IA = [sort(Irand[1:int(n)]) for n in [M, M*0.1, M*0.01, M*0.001, M*0.0001]];
+    debug = false
+
+    if debug
+        println("row sizes: $([int(nnz(S)/S.n) for S in SA])");
+        println("I sizes: $([length(I) for I in IA])");
+        @printf("    S    |    I    | binary S | binary I |  linear  | best\n")
+    end
+
+    J = Jrand;
+    for I in IA
+        for S in SA
+            res = Any[1,2,3]
+            times = Float64[0,0,0]
+            best = [typemax(Float64), 0]
+            for searchtype in [0, 1, 2]
+                tres = @timed test_getindex_algs(S, I, J, searchtype)
+                res[searchtype+1] = tres[1]
+                times[searchtype+1] = tres[2]
+                if best[1] > tres[2]
+                    best[1] = tres[2]
+                    best[2] = searchtype
+                end
+            end
+
+            if debug
+                @printf(" %7d | %7d | %4.2e | %4.2e | %4.2e | %s\n", int(nnz(S)/S.n), length(I), times[1], times[2], times[3],
+                            (0 == best[2]) ? "binary S" : (1 == best[2]) ? "binary I" : "linear")
+            end
+            @assert res[1] == res[2] == res[3]
+        end
+    end
+end